# perf(plugin-pnp): only load package data when needed