--- a/packages/plugin-pnp/sources/PnpLinker.ts
+++ b/packages/plugin-pnp/sources/PnpLinker.ts
@@ -129,22 +129,28 @@ export class PnpInstaller implements Installer {
       // We never need to unplug soft links, since we don't control them
       pkg.linkType !== LinkType.SOFT;
 
-    let customPackageData = this.customData.store.get(pkg.locatorHash);
-    if (typeof customPackageData === `undefined`) {
-      customPackageData = await extractCustomPackageData(pkg, fetchResult);
-      if (pkg.linkType === LinkType.HARD) {
-        this.customData.store.set(pkg.locatorHash, customPackageData);
+    let customPackageData: CustomPackageData | undefined;
+    let dependencyMeta: DependencyMeta | undefined;
+
+    if (mayNeedToBeBuilt || mayNeedToBeUnplugged) {
+      customPackageData = this.customData.store.get(pkg.locatorHash);
+
+      if (typeof customPackageData === `undefined`) {
+        customPackageData = await extractCustomPackageData(pkg, fetchResult);
+        if (pkg.linkType === LinkType.HARD) {
+          this.customData.store.set(pkg.locatorHash, customPackageData);
+        }
       }
+
+      dependencyMeta = this.opts.project.getDependencyMeta(pkg, pkg.version);
     }
 
-    const dependencyMeta = this.opts.project.getDependencyMeta(pkg, pkg.version);
-
     const buildScripts = mayNeedToBeBuilt
-      ? jsInstallUtils.extractBuildScripts(pkg, customPackageData, dependencyMeta, {configuration: this.opts.project.configuration, report: this.opts.report})
+      ? jsInstallUtils.extractBuildScripts(pkg, customPackageData!, dependencyMeta!, {configuration: this.opts.project.configuration, report: this.opts.report})
       : [];
 
     const packageFs = mayNeedToBeUnplugged
-      ? await this.unplugPackageIfNeeded(pkg, customPackageData, fetchResult, dependencyMeta)
+      ? await this.unplugPackageIfNeeded(pkg, customPackageData!, fetchResult, dependencyMeta!)
       : fetchResult.packageFs;
 
     if (ppath.isAbsolute(fetchResult.prefixPath))